thermal: Update THERMAL_MAX_TRIPS to 48 [NCT]
Also changed mask field from int to u64 to support the larger trip point
number.

Split the nct part of this change from the original change:
http://git-master/r/327982. This was done during transition to 4.4
kernel.

Reviewed-on: http://git-master/r/327982

+++ drivers/misc/nct1008.c
@@ -1185,7 +1185,7 @@ static int nct1008_probe(struct i2c_client *client,
 	struct nct1008_data *data;
 	int err;
 	int i;
-	int mask = 0;
+	u64 mask = 0;
 	char nct_int_name[THERMAL_NAME_LENGTH];
 	char nct_ext_name[THERMAL_NAME_LENGTH];
 
@@ -1242,7 +1242,7 @@ static int nct1008_probe(struct i2c_client *client,
 	for (i = 0; i < data->plat_data.num_trips; i++)
 		if (strcmp(data->plat_data.trips[i].cdev_type,
 						"shutdown_warning"))
-			mask |= (1 << i);
+			mask |= (1ULL << i);
 
 	if (data->plat_data.loc_name) {
 		strcpy(nct_int_name, "Tboard_");
